+++ discord/member.py
@@ -296,6 +296,7 @@ class Member(discord.abc.Messageable, _UserTag):
         "_banner",
         "communication_disabled_until",
         "flags",
+        "_avatar_decoration",
     )
 
     if TYPE_CHECKING:
@@ -317,6 +318,7 @@ class Member(discord.abc.Messageable, _UserTag):
         communication_disabled_until: datetime.datetime | None
         primary_guild: PrimaryGuild | None
         collectibles: Collectibles | None
+        avatar_decoration: Asset | None
 
     def __init__(
         self, *, data: MemberWithUserPayload, guild: Guild, state: ConnectionState
@@ -341,6 +343,7 @@ class Member(discord.abc.Messageable, _UserTag):
             data.get("communication_disabled_until")
         )
         self.flags: MemberFlags = MemberFlags._from_value(data.get("flags", 0))
+        self._avatar_decoration: Asset | None = data.get("avatar_decoration_data")
 
     def __str__(self) -> str:
         return str(self._user)
@@ -418,6 +421,7 @@ class Member(discord.abc.Messageable, _UserTag):
         self._banner = member._banner
         self.communication_disabled_until = member.communication_disabled_until
         self.flags = member.flags
+        self._avatar_decoration = member._avatar_decoration
 
         # Reference will not be copied unless necessary by PRESENCE_UPDATE
         # See below
@@ -449,6 +453,7 @@ class Member(discord.abc.Messageable, _UserTag):
             data.get("communication_disabled_until")
         )
         self.flags = MemberFlags._from_value(data.get("flags", 0))
+        self._avatar_decoration = data.get("avatar_decoration_data")
 
     def _presence_update(
         self, data: PartialPresenceUpdate, user: UserPayload
@@ -509,6 +514,19 @@ class Member(discord.abc.Messageable, _UserTag):
             ) = modified
             # Signal to dispatch on_user_update
             return to_return, u
+
+    @property
+    def display_avatar_decoration(self) -> Asset | None:
+        """Returns the member's guild specific avatar decoration if one is set. Otherwise,
+        their global avatar decoration is returned.
+
+        .. versionadded:: 2.8
+        """
+        if self._avatar_decoration is None:
+            return self._user.avatar_decoration
+        return Asset._from_avatar_decoration(
+            self._state, self.id, self._avatar_decoration.get("asset")
+        )
 
     @property
     def status(self) -> Status: